+++ packages/react/src/NodeViewContent.tsx
@@ -5,12 +5,13 @@ export interface NodeViewContentProps {
   as?: React.ElementType,
 }
 
-export const NodeViewContent: React.FC<NodeViewContentProps> = props => {
+export const NodeViewContent: React.FC<NodeViewContentProps> = React.forwardRef((props, ref) => {
   const Tag = props.as || 'div'
 
   return (
     <Tag
       {...props}
+      ref={ref}
       data-node-view-content=""
       style={{
         ...props.style,
@@ -18,4 +19,4 @@ export const NodeViewContent: React.FC<NodeViewContentProps> = props => {
       }}
     />
   )
-}
+})