+++ packages/react/src/NodeViewWrapper.tsx
@@ -6,13 +6,14 @@ export interface NodeViewWrapperProps {
   as?: React.ElementType,
 }
 
-export const NodeViewWrapper: React.FC<NodeViewWrapperProps> = props => {
+export const NodeViewWrapper: React.FC<NodeViewWrapperProps> = React.forwardRef((props, ref) => {
   const { onDragStart } = useReactNodeView()
   const Tag = props.as || 'div'
 
   return (
     <Tag
       {...props}
+      ref={ref}
       data-node-view-wrapper=""
       onDragStart={onDragStart}
       style={{
@@ -21,4 +22,4 @@ export const NodeViewWrapper: React.FC<NodeViewWrapperProps> = props => {
       }}
     />
   )
-}
+})